Moved Settings tests to a directory
